Fix some compiler warnings

+++ include/aws/cryptosdk/private/cipher.h
@@ -18,6 +18,8 @@
 
 #include <aws/cryptosdk/header.h>
 #include <aws/common/byte_buf.h>
+#include <stdlib.h>
+#include <stdint.h>
 
 /**
  * Internal cryptographic helpers.
@@ -36,7 +38,7 @@ struct aws_cryptosdk_alg_properties {
      */
     const struct aws_cryptosdk_alg_impl *impl;
 
-    int data_key_len, content_key_len, iv_len, tag_len;
+    size_t data_key_len, content_key_len, iv_len, tag_len;
 
     enum aws_cryptosdk_alg_id alg_id;
 };
